+++ src/data/models/LeaderboardModel.hh
@@ -196,7 +196,7 @@ public:
         m_pCapturedValueDefinitionHits.Reset();
     }
 
-    static constexpr int MaxSerializedLength = 65535;
+    static constexpr size_t MaxSerializedLength = 65535;
 
 protected:
     void OnValueChanged(const IntModelProperty::ChangeArgs& args) override;
